+++ server.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const userRoutes = require("./src/user/routes");
+const cardBenefitsInfo = require("./src/cardbenefitsinfo/routes");
 
 const app = express();
 const port = 3000;
@@ -14,5 +15,6 @@ app.get("/", (req, res) => {
 });
 
 app.use("/api/v1/users", userRoutes);
+app.use("/api/v1/cardbenefitsinfo", cardBenefitsInfo);
 
 app.listen(port, () => console.log(`app listening on port ${port}`));
