+++ src/user/controller.js
@@ -1,5 +1,5 @@
 const pool = require('../../db');
-const queries = require('./queries');
+const queries = require('../user/queries');
 
 const getUsers = (req, res) => {
   pool.query(queries.getUsers, (error, results) => {
@@ -8,23 +8,6 @@ const getUsers = (req, res) => {
   });
 };
 
-const getCardBenefitsInfo = (req, res) => {
-  pool.query(queries.getCardBenefitsInfo, (error, results) => {
-    if (error) throw error;
-    res.status(200).json(results.rows);
-  });
-};
-
-const getCardCompany = (req, res) => {
-  const cardCompanyName = req.params.cardcompany;
-  pool.query(queries.getCardCompany, [cardCompanyName], (error, results) => {
-    if (error) throw error;
-    res.status(200).json(results.rows);
-  });
-};
-
 module.exports = {
   getUsers,
-  getCardBenefitsInfo,
-  getCardCompany,
 };
