+++ src/user/routes.js
@@ -1,10 +1,8 @@
 const { Router } = require("express");
-const controller = require("./controller");
+const controller = require("../user/controller");
 
 const router = Router();
 
 router.get("/", controller.getUsers);
-router.get("/get/cardinfo", controller.getCardBenefitsInfo);
-router.get("/get/cardcompany/:cardcompany", controller.getCardCompany);
 
 module.exports = router;
